requirements.txt
pytest
pyHamcrest
pytest-bdd
pywinauto